// lib/types.ts
export interface User {
    id?: string;
    email?: string;
    name?: string;
    phone?: string;
    address?: string;
    usertype?: number;
    shopid?: ShopBrief[];
    tobeowner?: boolean;
    token?: string;
    logintime?: number;
}

export interface UserLoginType {
    email: string;
    password: string;
}

export interface UserAddType {
    name: string;
    email: string;
    password: string;
    phone: string;
    address: string;
    usertype: number;
    tobeowner: boolean;
}

export interface PasswdCheck {
    mail_sent: number;
    numForCheck: string;
    token: string;
}

export interface UserContextType {
    isLoggedIn: boolean; 
    showShopGuide: boolean;
    user: User; 
    login: (user?: User) => void; 
    logout: () => void;
    indexpageswitch: () => void;
}

export interface OrdersContextType {
    orderlist: MealOrderType[];
    update: (orders?: MealOrderType[]) => void;
}
// ...
export interface ShopDataType{
    id: string;
    shopname: string;
    foodsupply: string;
    profileimage: string;
    owner: string; 
    onboard: boolean;
    blockusers?: string[];
    created?: string;
} 

export interface ShopBrief{
    id: string;
    shopname: string;
    onboard: boolean;
}
// ...
export interface MealOrderElm {
    id: string;
    profileimage?: string;
    mealname: string;
    unitprice: number;
    quantity?: number;
}
  
export interface MealOrderRecord {
    userId?: string;
    userName?: string;
    orderList: MealOrderElm[];
    sum: number;
    tax: number;
    address: string;
    active: boolean;
    orderstatus: number;
    statushistory: string[];
    created?: string;
}

export interface MealOrderType {
    id: string;
    shopId?: string;
    shopName?: string;
    userId?: string;
    userName?: string;
    orderList: MealOrderElm[];
    sum: number;
    tax: number;
    address: string;
    active: boolean;
    orderstatus: number;
    statushistory: string[];
    created?: string;
    publiccode?: string;
}

export interface CurrentOrders {
   logintime?: number; 
   userId?: string;
   orderList: MealOrderType[];
}

export interface OrderShopsElm {
    id: string;
    shopname?: string;
    profileimage?: string;
    count: number;
    cancel: number;
}